Fixed incorrect merge
Should just be exactly the same as before but with the code changes I have made actually merged into the repo.
