--- Erebus-Orion/world_gen/boxTemplate.txt
@@ -1,21 +1,0 @@
-    Transform {{
-        children [
-            DEF WALL{5} Solid {{
-            translation {0} 0.75 {1}
-            children [
-                Shape {{
-                appearance Appearance {{
-                    material Material {{
-                    diffuseColor 0.32549 0.32549 0.32549
-                    }}
-                }}
-                geometry DEF WALLBOX{5} Box {{
-                    size {2} 1.5 {3}
-                }}
-                }}
-            ]
-            name "{4}"
-            boundingObject USE WALLBOX{5}
-            }}
-        ]
-    }}
