--- Erebus-Orion/world_gen/fileHeader.txt
@@ -1,115 +1,0 @@
-#VRML_SIM R2019b utf8
-WorldInfo {
-}
-Viewpoint {
-  orientation -0.9999043393436948 0.012895774501650994 0.005001116040641551 0.740064875011294
-  position 0.0797377570748452 26.58021196874523 32.6072172027324
-}
-TexturedBackground {
-}
-TexturedBackgroundLight {
-}
-DEF WALLGROUP Group {
-  children [
-    Transform {
-      children [
-        Solid {
-          children [
-            Shape {
-              appearance Appearance {
-                material Material {
-                }
-              }
-              geometry DEF box Box {
-                size 25 0.1 20
-              }
-            }
-          ]
-          boundingObject USE box
-        }
-      ]
-    }
-    Transform {
-      children [
-        Solid {
-          translation 12.45 0.75 0
-          children [
-            Shape {
-              appearance Appearance {
-                material Material {
-                  diffuseColor 0.32549 0.32549 0.32549
-                }
-              }
-              geometry DEF box Box {
-                size 0.1 1.5 20
-              }
-            }
-          ]
-          name "solid(1)"
-          boundingObject USE box
-        }
-      ]
-    }
-    Transform {
-      children [
-        Solid {
-          translation -12.45 0.75 0
-          children [
-            Shape {
-              appearance Appearance {
-                material Material {
-                  diffuseColor 0.32549 0.32549 0.32549
-                }
-              }
-              geometry DEF box Box {
-                size 0.1 1.5 20
-              }
-            }
-          ]
-          name "solid(2)"
-          boundingObject USE box
-        }
-      ]
-    }
-    Transform {
-      children [
-        Solid {
-          translation 0 0.75 9.95
-          children [
-            Shape {
-              appearance Appearance {
-                material Material {
-                  diffuseColor 0.32549 0.32549 0.32549
-                }
-              }
-              geometry DEF box Box {
-                size 25 1.5 0.1
-              }
-            }
-          ]
-          name "solid(3)"
-          boundingObject USE box
-        }
-      ]
-    }
-    Transform {
-      children [
-        Solid {
-          translation 0 0.75 -9.95
-          children [
-            Shape {
-              appearance Appearance {
-                material Material {
-                  diffuseColor 0.32549 0.32549 0.32549
-                }
-              }
-              geometry DEF box Box {
-                size 25 1.5 0.1
-              }
-            }
-          ]
-          name "solid(4)"
-          boundingObject USE box
-        }
-      ]
-    }
