--- Erebus-Orion/world_gen/obstacleGroupTemplate.txt
@@ -1,5 +1,0 @@
-DEF OBSTACLEGROUP Group {{
-    children [
-{0}
-    ]
-}}
