+++ game/worlds/.newGeneratedWorld.wbproj
@@ -1,0 +1,13 @@
+Webots Project File version R2019b
+perspectives: 000000ff00000000fd0000000300000000000001ae000003d1fc0100000002fb0000001e00480074006d006c0052006f0062006f007400570069006e0064006f00770100000000000001ae0000006900fffffffb0000001a0044006f00630075006d0065006e0074006100740069006f006e0000000000ffffffff0000006900ffffff000000010000016c0000034cfc0200000001fb0000001400540065007800740045006400690074006f007201000000140000034c0000008500ffffff00000003000005d000000083fc0100000001fb0000000e0043006f006e0073006f006c006501000001b0000005d00000006900ffffff000004620000034c00000001000000020000000100000008fc00000000
+simulationViewPerspectives: 000000ff000000010000000200000100000005620100000002010000000100
+sceneTreePerspectives: 000000ff000000010000000200000282000000fa0100000002010000000200
+maximizedDockId: -1
+centralWidgetVisible: 1
+selectionDisabled: 0
+viewpointLocked: 0
+orthographicViewHeight: 1
+textFiles: 0 "worlds/newGeneratedWorld.wbt"
+robotWindow: robot
+renderingDevicePerspectives: robot0:camera;1;1;0;0
+renderingDevicePerspectives: robot1:camera;1;1;0;0
