+++ game/worlds/newGeneratedWorld.wbt
@@ -2,8 +2,8 @@
 WorldInfo {
 }
 Viewpoint {
-  orientation -0.9999043393436948 0.012895774501650994 0.005001116040641551 0.740064875011294
-  position 0.0797377570748452 26.58021196874523 32.6072172027324
+  orientation -0.9999311643478309 -0.010615345112831881 -0.004998101052066259 0.8800535361313244
+  position -0.2732641733022936 31.53950726490244 28.893394932267597
 }
 TexturedBackground {
 }
@@ -1438,6 +1438,7 @@ DEF OBSTACLEGROUP Group {
 }
 Robot {
   controller "supervisorTest2"
+  supervisor TRUE
   window "simulationControl1"
 }
 Robot {
