+++ world_gen/obstacleGroupTemplate.txt
@@ -1,0 +1,5 @@
+DEF OBSTACLEGROUP Group {{
+    children [
+{0}
+    ]
+}}
