+++ world_gen/obstacleTemplate.txt
@@ -1,0 +1,17 @@
+		DEF OBSTACLE{0} Solid {{
+            translation 0 -1000 0
+            children [
+                Shape {{
+                    appearance Appearance {{
+						material Material {{
+						diffuseColor 0.45 0.45 0.45
+						}}
+                    }}
+                    geometry DEF OBSTACLEBOX{0} Box {{
+						size {1} {2} {3}
+                    }}
+                }}
+            ]
+            name "obstacle{0}"
+            boundingObject USE OBSTACLEBOX
+        }}
